requirements.txt
streamlit==1.28.1
azure-identity==1.14.0
azure-storage-blob==12.18.0
azure-mgmt-datafactory==9.0.0
openai>=1.12.0,<2.0.0
python-dotenv==1.0.0
pandas==2.1.1
pyodbc==5.0.1
sqlalchemy==2.0.22
requests==2.31.0